refactor: move color and make jsdoc private
