+++ apis/nucleus/src/components/listbox/assets/styling.js
@@ -1,4 +1,4 @@
-import Color from './color';
+import Color from '@nebula.js/theme/src/color/color';
 
 const LIGHT = '#FFF';
 const DARK = '#000';
